% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.col.R
\name{barplot.col}
\alias{barplot.col}
\title{Barplot with error bars}
\usage{
\method{barplot}{col}(data, lwd = 1, length = 0.2, err = "ster",
  col = "grey", las = 1)
}
\arguments{
\item{data}{a data.frame. Non-numeric columns will be removed.}

\item{lwd}{line width of the error bars}

\item{length}{length of error bar heads}

\item{err}{error bars show standard error (ster) or standard deviation (sd)}
}
\description{
Returns a barplot based on a given data.frame. The mean value of each column will be calculated and used for the plot. Error bars will be added, using standard error (default) or standard deviation.
}
\examples{
x <- c(2,4,3,5,NA,5)
sterr(x)
}

