% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakPointsArrhenius.R
\name{breakPointsArrhenius}
\alias{breakPointsArrhenius}
\title{produce an Arrhenius plot for groups and calculate optimum temperature.}
\usage{
breakPointsArrhenius(x, group, col = "black", marks = T, obs = 3,
  error.bars = T, ylim = c(4, 5.2), pch = 16, legend = T,
  xlab = "Temperature (<U+00B0>C)", ylab = "ln(HR max)")
}
\arguments{
\item{x}{data matrix, numbers or NAs only, samples in lines, one column per measurement step.}

\item{col}{color vector. Should have the same length as number of groups. Otherwise grey shades are used.}

\item{marks}{add marks for the optimum temperature}

\item{obs}{minimum number of observations to calculate the correlation in the linear phase. Default is 3.}

\item{error.bars}{if TRUE, standard error bars are added}

\item{ylim}{define own ylim}

\item{pch}{define dot style}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{groups}{group vector (must have same length as rows in data)}

\item{legeng}{if TRUE, a legend is added}
}
\description{
Plots and calcualtes the breakpoints for optimum temperature according to the Arrhenius method. The function censecutively adds measurements and calculates the correlation coefficient. If the coefficient is not increasing anymore it assumes that the linear phase ended and calculates the break point.
}
\examples{
breakPointsArrhenius(x=data[,-1], group=groups)
}

