% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakPointsQ10.R
\name{breakPointsQ10}
\alias{breakPointsQ10}
\title{Plot Q10 values for groups and calculate optimum temperature.}
\usage{
breakPointsQ10(x, group, col = "black", marks = T, error.bars = T,
  ylim = c(0, 3), pch = 16, legend = T, threshold = 1.9,
  xlab = "start point temperature", ylab = "Q10 value")
}
\arguments{
\item{x}{data matrix, numbers or NAs only, samples in lines, one column per measurement step.}

\item{col}{color vector. Should have the same length as numberof groups. Otherwise grey shades are used.}

\item{marks}{add marks for the optimum temperature}

\item{error.bars}{if TRUE, standard error bars are added}

\item{ylim}{define own ylim}

\item{pch}{define dot style}

\item{threshold}{defines the threshold for optimum temperature. Default is 1.9}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{groups}{group vector (must have same length as rows in data)}

\item{legeng}{if TRUE, a legend is added}
}
\description{
Plots and calcualtes the breakpoints for optimum temperature according to Q10 values. The default threshold is 1.9. Optimum temperature is assumed at intersection of of this threshold and the Q10 value curve for the respective group.
}
\examples{
breakPointsQ10(x=data, group=groups, col=c("blue","red","black"))
}

