% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.plot.R
\name{HR.plot}
\alias{HR.plot}
\title{Plot heart rate data}
\usage{
HR.plot(data, groups, xlabels = colnames(data), dodge = 0.03, xlim = "",
  ylim = "", cex = 2, xlab = "degrees C",
  ylab = "maximum heart rate (bpm)", length = 0.02, lwd = 1,
  lwd.err = 1, main = "", las.x = 1, las.y = 2,
  legend.pos = "bottomright", legend = TRUE, bg.grid = T, col = "",
  pch = 16, at = "", cutoff = 0.5, sig.marks = "", leg.labels = "")
}
\arguments{
\item{data}{data matrix, numbers or NAs only, samples in lines, one column per measurement step.}

\item{groups}{group vector (must have same length as rows in data)}

\item{xlabels}{labels for the x-axis}

\item{dodge}{add a dodge on xaxis for groups}

\item{xlim}{define own xlim}

\item{ylim}{define own ylim}

\item{cex}{scaling of data points}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{length}{length of error bar heads}

\item{lwd}{line width of dot connectors}

\item{lwd.err}{line width of error bars}

\item{main}{title of plot}

\item{las.x}{text orientation for x-axis labels (1=horizontal, 2=vertical)}

\item{las.y}{text orientation for y-axis (2=horizontal, 1=vertical)}

\item{legend.pos}{position of legend}

\item{legend}{if TRUE, a legend will be added to the plot}

\item{bg.grid}{if TRUE, a background grid will be drawn}

\item{col}{optinal color vector, default is grey scales}

\item{pch}{defines dot style. Can be a vector with length = number of groups}

\item{at}{define order of the groups and positioning for the legend labels}

\item{cutoff}{percentage of samples needed for draing a solid line, less samples will be drawn with a dashed line.}

\item{sig.marks}{vector of x-coordinates. Puts grey rectangle and asterisks in the background of the plot at the indicated position}

\item{leg.labels}{vector containing the labels for the legend}
}
\description{
Specialized plot for heart rate data. Every data point is plotted as a dot with standard error bars. Dots are connected by solid line but as soon as the number of samples represented by the dat gets below the cutoff, the connective line is dashed.
}
\examples{
HR.plot(data, groups)
}

