% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.err.R
\name{plot.err}
\alias{plot.err}
\title{Plot means and error bars}
\usage{
\method{plot}{err}(data, groups, at = FALSE, type = "dot", barwidth = 0.2,
  col = "black", err.bars = "sterr", cex = 2, xlab = "",
  ylab = "means", length = 0.1, lwd = 1, main = "", las.x = 1,
  las.y = 2, xaxt = T, cex.yaxis = 1, cex.xaxis = 1, yscale = c(1, 1),
  stats = T, stats.mark = T, sig.groups = F)
}
\arguments{
\item{data}{a numeric vector containing the data to plot.}

\item{groups}{vector containing the group information for data.}

\item{at}{optinal vector to arrange the position of the data points on the x-axis.}

\item{type}{type of plot: Either "dot" for dots, "bar" for bars or "diffbar" for bars, starting at the average.}

\item{barwidth}{if type="bar", barwidth defines the width of the bars.}

\item{col}{vector defining the color for each group. Must have the same length as number of groups.}

\item{err.bars}{error bars show standard error ("sterr") or standard deviation ("sd").}

\item{cex}{defines the size of the dots.}

\item{xlab}{label for x-axis.}

\item{ylab}{label for y-axis.}

\item{length}{defines the length of the arrow heads of the error bars}

\item{lwd}{line width of the error bars.}

\item{main}{title of the plot}

\item{las.x}{text orientation of the x-axis labels.}

\item{las.y}{text orientation of the y-axis labels.}

\item{xaxt}{add x-axis if TRUE}

\item{cex.yaxis}{scaling for y-axis labels}

\item{cex.xaxis}{scaling for x-axis labels}

\item{yscale}{vector of length two to scale the ylim's upper and lower limit}

\item{stats}{add ANOVA results to the plot}

\item{stats.mark}{add stars to significant ANOVA results}

\item{sig.groups}{vector with significant marker, which will be places over the error bars (order is 1:n and not according to at!)}
}
\description{
Plot the mean of a data vector, grouped be the group vector and add error bars.
}
\examples{
plot.err(data=c(1,2,3,5,3,2,4,12), groups = c(rep("a",4),rep("b",4)))
}

