% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.err.two.factor.R
\name{plot.err.two.factor}
\alias{plot.err.two.factor}
\title{Plot means and error in group and subgroup}
\usage{
\method{plot}{err.two.factor}(data, group1, group2, rect.width = c(-0.3, 0.3),
  at = FALSE, colG1 = "grey", colG2 = "black", err.bars = "sterr",
  cex = 2, shading = FALSE, xlab = "", ylab = "means", length = 0.1,
  lwd = 1, main = "", las.x = 1, las.y = 2, xaxt = T,
  legend.pos = "topright", cex.yaxis = 1, cex.xaxis = 1, yscale = c(1,
  1))
}
\arguments{
\item{data}{a numeric vector containing the data to plot.}

\item{group1}{vector containing the (background) group information for data.}

\item{group2}{vector containing the sub-group information for data.}

\item{rect.width}{width of the rectangle of group1}

\item{at}{optional vector to arrange the position of the data points on the x-axis. Must be integer.}

\item{colG1}{vector defining the colors for group1. Must have the same length as number of groups.}

\item{colG2}{vector defining the colors for group2. Must have the same length as number of groups.}

\item{err.bars}{error bars show standard error ("sterr") or standard deviation ("sd").}

\item{cex}{defines the size of the dots (means of group2).}

\item{shading}{intensity (between 1 and 99) of overlay of the background (group1) with a transparent white layer (default FALSE)}

\item{xlab}{label for x-axis.}

\item{ylab}{label for y-axis.}

\item{length}{defines the length of the arrow heads of the error bars}

\item{lwd}{line width of mean marker (group1) and the error bars (group2).}

\item{main}{title of the plot}

\item{las.x}{text orientation of the x-axis labels.}

\item{las.y}{text orientation of the y-axis labels.}

\item{xaxt}{add x-axis if TRUE}

\item{legend.pos}{postion of legend (for group2)}

\item{cex.yaxis}{scaling for y-axis labels}

\item{cex.xaxis}{scaling for x-axis labels}

\item{yscale}{vector of length two to scale the ylim's upper and lower limit}
}
\description{
Plot group data in the background and subgroups in the foreground with mean+/-sterr
}
\examples{
plot.err.two.factor(dat$V, dat$Gr.navn, dat$feed, colG1=cols[3:1], colG2 = cols[6:4], lwd=3, shading=30, at=c(3,2,1))
}

