% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q10.R
\name{q10}
\alias{q10}
\title{Calculate Q10 values column-wise.}
\usage{
q10(x)
}
\arguments{
\item{x}{numeric matrix containing the consecutive measurements in columns and test subjects in rows.}
}
\description{
Calculates Q10 values for an input matrix column-wise according to the formula (R2/R1)^10 with R1 and R2 being consecutive measurements.
}
\examples{
q10(matrix(11:20, ncol=5))
}

