% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sterr.R
\name{sterr}
\alias{sterr}
\title{Calculate the standard error.}
\usage{
sterr(x)
}
\arguments{
\item{x}{numeric vector. NAs are allowed but will be removed before calculation.}
}
\description{
Calculates the standard error of numbers in a vector. NAs are removed.
}
\examples{
x <- c(2,4,3,5,NA,5)
sterr(x)
}

