% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGC_calculation.R
\name{TGC_calculation}
\alias{TGC_calculation}
\title{Calculate the Thermal Growth Coefficent.}
\usage{
TGC_calculation(startWeight, finalWeight, days, degrees = 12)
}
\arguments{
\item{startWeight}{numeric value, weight (g) at first measurement}

\item{finalWeight}{numeric value, weight (g) at second measurement}

\item{days}{numeric value, days between the two measurements}

\item{degrees}{numeric value, average temperature of the water}
}
\description{
Calculates the Thermal Growth Coefficent (TGC) based on initial weight (grams), final weight (grams), number of days between the two weight measurements and the temperature (degrees C).
}
\examples{
TGC_calculation(1.7, 24.8, 65, 12)
}

